Conflicts:
	GameCardsDev/screens/GamePlayScreen.cpp

+++ GameCardsDev/UI/Widgets/MobImage.cpp
@@ -30,10 +30,10 @@ bool MobImage::statContains(int x, int y, int width, int height, int pointX, int
 	Rect *r = NULL;
 	switch(orientation) {
 		case LANDSCAPE:
-			r = new Rect((this->getWidth()-((y + height)*((double)this->getWidth()/350))),
-					x*((double)this->getHeight()/250),
-					height*((double)this->getWidth()/350),
-					width*((double)this->getHeight()/250));
+			r = new Rect((int)((this->getWidth()-((y + height)*((double)this->getWidth()/350)))),
+					(int)(x*((double)this->getHeight()/250)),
+					(int)(height*((double)this->getWidth()/350)),
+					(int)(width*((double)this->getHeight()/250)));
 			break;
 		case PORTRAIT:
 			r = new Rect((x*this->getWidth()/250),(y*this->getHeight()/350),width*this->getWidth()/250,height*this->getHeight()/350);
